refactor(ui): remove storybook specific styling
