+++ packages/ui/src/design-system/side-drawer/side-drawer-content.css.ts
@@ -1,4 +1,4 @@
-import { sx, style, keyframes, globalStyle } from '../../design-tokens';
+import { sx, style, keyframes } from '../../design-tokens';
 
 const slideIn = keyframes({
   from: {
@@ -20,11 +20,11 @@ const slideOut = keyframes({
 
 export const container = style([
   sx({
-    width: '$dialog',
-    height: '$auto',
     top: '$24',
     bottom: '$24',
     right: '$24',
+    width: '$dialog',
+    height: '$auto',
   }),
   {
     position: 'fixed',
@@ -39,34 +39,3 @@ export const container = style([
     },
   },
 ]);
-
-export const content = style([
-  sx({
-    boxShadow: '$dialog',
-    borderRadius: '$extraLarge',
-    backgroundColor: '$side_drawer_container_bgColor',
-    width: '$fill',
-    height: '$fill',
-    display: 'grid',
-  }),
-  {
-    gridTemplateAreas: `
-    "header"
-    "body"
-    "footer"`,
-    gridTemplateRows: 'auto 1fr auto',
-    overflow: 'hidden',
-  },
-]);
-
-export const storybook = style({});
-
-// storybook only
-globalStyle(`#storybook ${storybook}`, {
-  position: 'relative',
-  top: 'auto',
-  bottom: 'auto',
-  right: 'auto',
-  animation: 'none',
-  height: '100%',
-});
